add basic backup stats readout

+++ pkg/backup/chunk_packer.go
@@ -86,7 +86,7 @@ func (cp *chunkPacker) Complete() (isJournalComplete bool) {
 
 		// Upload chunk (and run unrelated parallel func during upload)
 		objName := "chunks/" + chunkName
-		cp.vlog.Printf("chunkPacker: Complete: writing object '%s' to cloud (%s)", objName, util.FormatBytesAsString(len(ciphertextChunkBuf)))
+		cp.vlog.Printf("chunkPacker: Complete: writing object '%s' to cloud (%s)", objName, util.FormatBytesAsString(int64(len(ciphertextChunkBuf))))
 		err = cp.objst.UploadObjFromBuffer(cp.ctx, cp.bucket, objName, ciphertextChunkBuf, objstore.ComputeETag(ciphertextChunkBuf))
 		if err != nil {
 			log.Printf("error: chunkPacker.Complete: failed while uploading '%s': %v\n", chunkName, err)
